requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.12.1+cu116 
torchvision==0.13.1+cu116
pytorch-lightning==1.7.7
tqdm
wandb
PyYAML==6.0
git+https://github.com/lucasb-eyer/pydensecrf.git
mmcv-full==1.6.2 -f https://download.openmmlab.com/mmcv/dist/cu116/torch1.12.0/index.html
mmsegmentation==0.28.0
flow_vis
scipy
scikit-image
